refactor: extract summary form
