+++ src/app/(main)/summary/_components/summaryFormCard/index.ts
@@ -1,3 +1,4 @@
 export { SummaryFormCard } from './SummaryFormCard'
 
 export { SummaryFormHeader } from './SummaryFormHeader'
+export { SummaryForm } from './SummaryForm'
