+++ src/services/summary/index.ts
@@ -1,2 +1,2 @@
-export type { SummaryData, SummaryResponse, SummaryForm } from './types'
+export type { SummaryData, SummaryResponse, SummaryFormData } from './types'
 export { addSummary, getSummary } from './api'
